# frontend/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI焦化厂智能化平台 API",
    "description": "智能配煤 / 焦炉加热控制 / 设备PdM / 安全视觉 四大应用后端",
    "version": "1.1.0"
  },
  "paths": {
...
    "/pdm/devices": {
      "get": {
        "tags": [
          "设备PdM"
        ],
        "summary": "监控设备列表",
        "description": "返回监控设备清单，支持按优先级P0/P1/P2过滤。",
        "operationId": "list_devices_pdm_devices_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^P[012]$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevicesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/pdm/alarms": {
      "get": {
        "tags": [
          "设备PdM"
        ],
        "summary": "PdM告警列表",
        "description": "查询两级预警告警（4.3.2节），响应契约见 docs/API文档.md §5。\n\n数据源：PostgreSQL pdm_alarm 表（方案§3.4.6）；level 为严重度，\n两级分类在 source 字段（level1_anomaly 实时异常 / level2_trend_forecast 趋势分级）。",
        "operationId": "list_alarms_pdm_alarms_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "equipment_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment Id"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^(WARNING|DANGER)$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "acknowledged",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Acknowledged"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdmAlarmsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vision/alarms": {
      "get": {
        "tags": [
          "安全视觉"
        ],
        "summary": "视觉告警列表",
        "description": "查询视觉告警（vision_alarm 表，方案§3.4.7），按时间倒序。",
        "operationId": "list_alarms_vision_alarms_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scene",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "场景过滤: helmet/fire/intrusion/gas_leak/gauge/coke_cake",
              "title": "Scene"
            },
            "description": "场景过滤: helmet/fire/intrusion/gas_leak/gauge/coke_cake"
          },
          {
            "name": "area",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "区域过滤（如 焦炉炉顶）",
              "title": "Area"
            },
            "description": "区域过滤（如 焦炉炉顶）"
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "acknowledged",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "是否已确认",
              "title": "Acknowledged"
            },
            "description": "是否已确认"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisionAlarmsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/vision/alarms/{alarm_id}/ack": {
      "post": {
        "tags": [
          "安全视觉"
        ],
        "summary": "视觉告警确认",
        "description": "确认告警：记录处理人与处置意见；误报标记回流训练集迭代（每周重训消费）。",
        "operationId": "ack_alarm_vision_alarms__alarm_id__ack_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "alarm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Alarm Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AckRequest": {
        "properties": {
          "handler": {
            "type": "string",
            "title": "Handler",
            "description": "处理人"
          },
          "comment": {
            "type": "string",
            "title": "Comment",
            "description": "处置意见",
            "default": ""
          },
          "is_false_positive": {
            "type": "boolean",
            "title": "Is False Positive",
            "description": "是否误报（回流用于模型迭代）",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "handler"
        ],
        "title": "AckRequest",
        "description": "告警确认请求。"
      },
      "AckResponse": {
        "properties": {
          "alarm_id": {
            "type": "integer",
            "title": "Alarm Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "alarm_id",
          "status"
        ],
        "title": "AckResponse",
        "description": "POST /vision/alarms/{id}/ack 响应。"
      },
...
      "DevicesResponse": {
        "properties": {
          "devices": {
            "items": {
              "$ref": "#/components/schemas/PdmDevice"
            },
            "type": "array",
            "title": "Devices"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "devices",
          "total"
        ],
        "title": "DevicesResponse",
        "description": "GET /pdm/devices 响应。"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PdmAlarm": {
        "properties": {
          "alarm_id": {
            "type": "integer",
            "title": "Alarm Id"
          },
          "equipment_id": {
            "type": "integer",
            "title": "Equipment Id"
          },
          "level": {
            "type": "string",
            "title": "Level",
            "description": "告警级别：WARNING/DANGER"
          },
          "source": {
            "type": "string",
            "title": "Source",
            "description": "来源：level1_anomaly/level2_trend_forecast"
          },
          "metric": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric"
          },
          "metric_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metric Value"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          },
          "iso10816_zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iso10816 Zone"
          },
          "msg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msg"
          },
          "acknowledged": {
            "type": "boolean",
            "title": "Acknowledged"
          },
          "handler": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Handler"
          },
          "ack_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ack Comment"
          },
          "acked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acked At"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "alarm_id",
          "equipment_id",
          "level",
          "source",
          "acknowledged",
          "ts"
        ],
        "title": "PdmAlarm",
        "description": "PdM 告警（GET /pdm/alarms 响应元素，方案§4.3.2 两级预警）。"
      },
      "PdmAlarmsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PdmAlarm"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PdmAlarmsResponse",
        "description": "GET /pdm/alarms 响应。"
      },
      "PdmDevice": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "priority": {
            "type": "string",
            "title": "Priority",
            "description": "优先级：P0/P1/P2"
          },
          "sensors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sensors",
            "description": "监测传感器类型"
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "name",
          "priority",
          "sensors"
        ],
        "title": "PdmDevice",
        "description": "监控设备（GET /pdm/devices 响应元素，方案§4.3.1 监控对象清单）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VisionAlarm": {
        "properties": {
          "alarm_id": {
            "type": "integer",
            "title": "Alarm Id"
          },
          "scene": {
            "type": "string",
            "title": "Scene"
          },
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "camera_id": {
            "type": "string",
            "title": "Camera Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "msg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Msg"
          },
          "snapshot_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Url"
          },
          "acknowledged": {
            "type": "boolean",
            "title": "Acknowledged"
          },
          "is_false_positive": {
            "type": "boolean",
            "title": "Is False Positive"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          }
        },
        "type": "object",
        "required": [
          "alarm_id",
          "scene",
          "camera_id",
          "acknowledged",
          "is_false_positive",
          "ts"
        ],
        "title": "VisionAlarm",
        "description": "视觉告警（GET /vision/alarms 响应元素，方案§3.4.7/§4.4.1）。"
      },
      "VisionAlarmsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VisionAlarm"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "VisionAlarmsResponse",
        "description": "GET /vision/alarms 响应。"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}